Update patch set 6
Patch Set 6:

> Hi Tom,
 > 
 > I took all your comments and have pushed a fix for them. However, I
 > wasn't able to test the multiline bug as when I try to drag and
 > drop a link that isn't from my browser, Eclipse doesn't take it as
 > a drag object. Hopefully you could run a quick test to see if it is
 > fixed and if not, perhaps you can let me know how you have been
 > testing that?

When the clipboard contains multi-line text that starts with a valid git URL on the first line and the git clone wizard is then opened normally (not via DND), we end up with the full multi-line text in the text field.

Anyway, since I found it faster to fix this and clean up the remaining issues than writing yet another review I've gone ahead and made the fixes.

Patch-set: 6
Label: Verified=0