Update patch set 3
Patch Set 3:

Thanks for the review Andrey! I moved the complete dialog creation to a new method (showCheckoutConfirmationDialog) and described the behavior in the doc (based on the used jface version). Are you ok with this or should  the new part put into a separate method only?

Patch-set: 3
Label: Verified=0