Update patch set 4
Patch Set 4:

> @David, see comment from Thomas:
 > As long as we build against Neon on Hudson this will not compile.
 > 
 > -> The EGit team has to decide, if they want to compile against
 > Oxygen.

My 2¢: I'm rather against that. In fact, I'd rather we build against an even older target (Mars, which is our oldest target). EGit is supposed to work on Luna. So in theory we should build against a Luna target (but we don't have one anymore) and run the tests on Oxygen, or build against a newer target such as Oxygen but run the tests on Luna (though that may or may not uncover such problems, depending on whether the code in question is exercised in tests, so we would have to be much more stringent with requiring tests for any new code).

Already in Neon there were changes that can get you in trouble when writing new code (an example: default methods in interfaces, for instance IContentProvider.inputChanged(): if you create a new content provider and forget to override this method, the code will crash on Mars and earlier with "cannot call abstract method", but if you develop, build, and test against a Neon or newer target, you'll never notice that until it's too late, and JDT will not prompt you to override this method since it's a default method as of Neon). In Oxygen, there've been many more such changes, and new methods (functional interface helper methods etc.), so chances will be even higher that you write code that won't work on anything pre-Oxygen and that you won't notice it.

I also think that all this version checking and doing different things based on the JFace version is overkill for something as trivial as changing an "Ok" button label to "Check out".

That said, I think the decision what platform we build against is ultimately Matthias' decision.

Patch-set: 4