Update patch set 4
Patch Set 4:

> > @David, see comment from Thomas:
 > > As long as we build against Neon on Hudson this will not compile.
 > >
 > > -> The EGit team has to decide, if they want to compile against
 > > Oxygen.
 > 
 > My 2¢: I'm rather against that.

I fully agree with Thomas. Building against Oxygen is a no-go as long as we support older platforms, and we do that and I don't think we are going to change it any time soon. We always tried to support older releases as long as possible, for the cost of not using "latest greatest" stuff from the platform.

Patch-set: 4
Reviewer: Gerrit User 1179 <1179@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0