CommitEditor: fix check-out button display
The button would show disabled when the editor was not active. That's
correct. But when clicked, editor activation would occur first, and
then the command was executed. The net user-visible effect was that
we had a "disabled" button in the UI that was armed and would perform
a command all the same.

That's not good. Resolve similar to the PDE ManifestEditor's launch
bar actions: show them always as enabled. This is achieved by not
contributing the commands directly to the toolbar but using an
intermediary action.

Also use only the current selection in the handler's activation
expression.

Bug: 516679
Change-Id: Ibce2adbae1d6bf70009c1b9fabbd285d2f21f4bb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>