Remove defaultHandlers for CommitEditor commands
Provide dedicated handlers with an appropriate activation expression
instead. The commands work for one or several RepositoryCommits. The
expression must not refer to the active part since these commands can
be invoked from either the CommitEditor or the GitHistoryPage.

In the CommitEditor, use intermediary actions for the toolbar
contributions to avoid enablement problems (see bug 516679).

Bug: 495064
Change-Id: I0297a4ef6cf37b1662181570fad9a1fdecaed33d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>