Add ICommitMessageProvider2 for caret positioning in commit messages
The caret in the CommitMessageComponent is placed, according to the
value of getDesiredCaretPosition() of the first implementation of
ICommitMessageProvider2, that provides caret position. In case there are
multiple implementations available, which provide a defined caret
position, they are ignored.
Commit messages are not trimmed anymore, to ensure consistency between a
provided message and the provided caret position.

Bug: 516867
Change-Id: I1b0ffd33ef37196d53077d09774672c2b033835f
Signed-off-by: Stefan Rademacher <stefan.rademacher@tk.de>