Add ICommitMessageProvider2 for caret positioning in commit messages
The caret in the CommitMessageComponent is placed, according to the
value of getCaretPosition() of the first implementation of
ICommitMessageProvider2, that was discovered. In case there are multiple
implementations available, they are ignored.
Commit messages are not trimmed anymore, to ensure consistency between a
provided message and the provided caret position.

Bug: 516867
Change-Id: I1b0ffd33ef37196d53077d09774672c2b033835f
Signed-off-by: Stefan Rademacher <stefan.rademacher@tk.de>