Contextual actions in tooltips in staging view
If there are selected elements in the staged/unstaged viewers,
provide quick access to the "Add to index"/"Remove from index"
actions via a toolbar shown in a tooltip when the mouse pointer
hovers over a selected element.

Also add those actions to the toolbars of the viewers, properly
enabled/disabled depending on whether there is a selection in
the viewer.

Change the unstage icon: "stage" is a green plus sign, but "unstage"
was an inverted "reset" icon. A red minus sign seems to be more
appropriate and less confusing.

Finally, add tooltips to the section headers of the staged and
unstaged viewers with explanations that should make it easier
for people new to git to get what this whole view is about.
(The notion of being able to compose a commit in the git way is
a novelty for many people new to git.)

Bug: 516969
Change-Id: I53f7d460aee079db25b78afd5a818bc7075547ce
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>