Provide option to copy file names in staging view to clipboard
Make the path of selected files and folders in the Staging view
copyable to the clipboard.
If multiple files are selected all paths are copied and the
individual paths are separated with System#lineSeparator.
No line separator is inserted after the last item.
This mirrors the behavior of the file list in the History view.

Bug: 510945
Change-Id: I75b490aa5219236d5a645ad122184a4a6b29492a
Signed-off-by: Fabian Pfaff <fabian.pfaff@vogella.com>