Update patch set 6
Patch Set 6:

(10 comments)

> How can I dynamically add a plugin (-> ICommitMessageProvider2) to a test runtime?

That's going to be difficult without injection and without mocking. Probably the best you can do is provide one statically in the test bundle via an executable extension factory that itself delegates to some singleton factory that creates the provider. That singleton factory can be known to the test class, and the test class can enable it just for its tests and disable it afterwards. The singleton factory would return a no-op commit message provider if disabled, and a test provider that does implement ICommitMessageProvider2 when enabled.

Just an idea; perhaps there's an easier way, but I can't think of one off the top of my head.

Patch-set: 6
Label: Verified=0