Update patch set 4
Patch Set 4:

(1 comment)

Hi Thomas,

Sorry for the delay. I talked to Stefan and he agrees that for our case (Papyrus) the solution works either way. So it is hard to tell which should be the correct behavior. Nevertheless, I propose to leave the order as it is now to keep this change less invasive.

If we come across a case that clearly favors one case over the other, we can adapt the behavior in a separate change. In such a change we would probably also need to ensure that there are no other areas in EGit where such a precedence assumption has been made with regard to Resource vs ResourceMapping.

What do you think?

Patch-set: 4
Reviewer: Gerrit User 79641 <79641@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0