Update patch set 20
Patch Set 20: Code-Review+1

(3 comments)

> INSTANCE does not yet accept a list.
 > Just wanted to know, if I added the second provider within the same
 > <extension> element according to your idea.

Yes, that was the idea.

I've described my idea with a list in INSTANCE in in-line comments. However, your approach also works and is probably good enough for a test for now, so unless you absolutely want to, there's no need to change the code again as far as I'm concerned.

Let's see what Matthias thinks. Code looks fine for me; but we'll have to file a CQ since it's more than 1000 changed lines.

Patch-set: 20