FetchGerritChangePage: also try to determine the patch set number
If we can determine a patch set number, try to build and insert a
valid change ref in the Change field instead of just the change
number. If the user knows it's valid, he doesn't have to wait for
the background ls-remote to finish and can continue right away. If
the Gerrit server does not actually have a change for the ref, the
fetch will then fail, of course.

Bug: 518492
Change-Id: Iddf58173a4272e5d8ef642827372c630184c5370
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>