Update patch set 2
Patch Set 2:

> Yes, we can show different button labels based on the rebase
 > result. Do you have a suggestion? Maybe simply "Close and Continue"
 > for the case you don't only close the dialog?

How about IDialogConstants.PROCEED_LABEL if we're in the conflict case and the user did not select "do nothing", and IDialogConstants.CLOSE_LABEL otherwise?

(BTW, that might also be a solution for the FetchFromGerritPage: use "Proceed" unless the user had chosen "update FETCH_HEAD only", in which case we'd show "Fetch").

Patch-set: 2