Correct sync info calculation for folders
The folder handling in the git "Team synchonization" support was wrong
since commit f46237a (bug 516426). That commit considered folders as
supervised -- which is correct since otherwise some changes may be
missed. It also tried to compensate for folders without changes showing
up in the synchronize view by setting their sync state to IN_SYNC if
they had only IN_SYNC descendants.

That broke the display of "compacted folders" in the synchronize view.
If a file is not IN_SYNC, the synchronize view is able to handle its
parent containers fine without needing any SyncInfo for them. When we
do supply sync info, the parent containers will be considered changed
themselves, and will show up explicitly in the synchronize view.

Therefore, simply return IN_SYNC for folders. The state recorded in the
GitSyncObjCache is bogus anyway.

Bug: 518646
Change-Id: I2970a50b2d11403b84dc1b9a000021f61ca92f90
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>