Update patch set 1
Patch Set 1:

There's three things I don't quite like about adding the menu to the form header:

1. it forces us to use an internal class (FormHeading).
2. we end up with identical updateHeadMeanu() methods in ReflogView and StagingView, but they're so specific that it makes no sense to factor them out and unify them.
3. it needs a slightly ugly hack to get the triangle appear initially.

Putting the menu into the normal view menu, as I did on the history page, seems a better choice for the point of view of the implementation. For a usability point of view, the menus in the form headers are sightly more noticeable.

Patch-set: 1