Work-around: don't auto-refresh ignored top-level folders in projects
The auto-refresh done by EGit on index changes until now refreshed
everything in all projects in a certain repository. As discussed in
bug 500106, this may be a very costly operation and may lock the
workspace for extended periods. Unfortunately, there doesn't seem
to be a reasonably simple way to get precise information about which
files in the workspace were touched by the last git command as was
suggested in bug 416199.

As a work-around, change the project refresh logic to exclude
git-ignored top-level folders in projects. That at least will
resolve the problem for the occurrence reported in the bug
(Node.js nodemodules folder).

Unchanged in this commit but IMO also questionable: refreshing the
projects runs with the workspace root as scheduling rule. This might
not be necessary since the actual refresh acquires per-project
scheduling rules anyway. Moreover, all projects are refreshed
sequentially. I wonder if this could not be parallelized and be
run without a workspace lock.

Bug: 500106
Change-Id: Ie301964fe2cbd9ec39f5d2146b84e04cfe8e0b82
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>