Update patch set 1
Patch Set 1:

> IMHO the best approche would be a drop-down box (Combo) beside the
 > filter box. This way its easy to discover and you can easily
 > identify the current repo and branch via the selected entry. And
 > you could remove the form header label, which results in less used
 > space.

A combo box in the toolbar will result in a rather wide view toolbar that will then often wrap and equally consume space. Also, it's probably not a good idea for the history view, where we have that dynamic search widget in the toolbar. In the reflog view, there's also the branch selector in the form heading on the right.

I think getting rid of the form headers is a separate issue (and if done, should be done for all git views: staging, reflog, rebase interactive so that we still have a uniform visual appearance). One possibility might be simply to use  the view description, which can be seen on the history view. It's much smaller... but also doesn't look that nice.

For giving access to repository switching (and branch switching in the reflog view) without using the form heading we have basically two options:

* new toolbar buttons with drop-down menus, or
* sub-menus in the view menu.

Patch-set: 1