Update patch set 1
Patch Set 1:

> try the following using a jgit clone:
 > * copy a change URL to the clipboard e.g. https://git.eclipse.org/r/#/c/99874/
 > * click "Fetch from Gerrit"
 > * wait until highest patch set number is auto-supplied
 > * now try to edit the patch set number in the change edit field by
 > hitting the delete key and try to enter 2
 > * the highest patch set number is supplied again so the edit is
 > lost
 
Hmmm. Will have a look.

 > when playing with this change I discovered that it seems we lost a
 > feature in this wizard: AFAIR the popup shown when the change field
 > is populated with a patch set used to show the commit message of
 > the selected patch set but currently it's always empty so this
 > popup is pretty useless

You mean the popup on the content assist dropdown? That always has said only "Patch set X of change Y". ("Always" meaning "for as long as I can remember" :-) ) We do not have the commit message. The dialog does a ls-remote; that doesn't return the commits. If we wanted to get more info, we would need to use an ssh or http REST endpoint to get info about open changes.

Patch-set: 1
Label: Verified=0