Update patch set 1
Patch Set 1:

> we use wikitext to update the documentation in egit.doc from our
 > wiki pages [1]
 > (user guide, contributor guide, new and noteworthy pages) so please
 > do not remove this but add the new repository.
 > 
 > Maybe we can remove the mylyn latest repository since the release
 > train repository already has mylyn ?
 > 
 > [1] https://wiki.eclipse.org/EGit/Contributor_Guide#EGit_3

Wasn't aware of that. So we need wikitext to build, but we don't really need it in the target. Our product (EGit/JGit) does not depend on it, only our toolchain. Can't we supply wikitext to the toolchain via some other means? Maven build dependency or some such? I'll have to see how exactly we do this doc generation...

But as a quick fix (and certainly for stable-4.8) I can simply add the new wikitext repo. Which do you prefer: the wikitext 3.0.6 referenced in mylyn.aggrcon, or snapshots/nightly, or the current highest available release, 3.0.9?

Patch-set: 1
Label: Verified=0