Update patch set 1
Patch Set 1:

> * repository sorting in the menu doesn't match the sorting we use
 > in repositories view, there we sort case insensitive
 > * when the repositories view is configured to follow selection and
 > I select another repository e.g. in history view I'd expect the
 > repositories view to also switch to the same repository

That is, with the current design of the history page, outright impossible. The repository is not part of the selection provided by that page.

I am working on a refactoring of the history view actions (getting rid of default handlers). As part of that, I'm also changing the selection provider of the history page. Once that's done, the repositories view should then be able to update when the repo selection changes in the history view. But similar problems exist for the staging view, the reflog view, and the rebase interactive view.

 > * can we show the repository selection menu in history view in the
 > same place where it's placed in staging and reflog views ? I think
 > this would be more consistent and requires one click less. The
 > selected project/folder/file path currently displayed top left
 > could be moved to the top right corner of the view

How? This a view description; I'm not aware of us having any control over it except setting the text.

 > * how about adding this menu also to the interactive rebase view ?

Sure.

Any preference about the mechanism? Toolbar button, view menu, or this form header menu hack?

The form headers in these view do take up quite some space, which is scarce in the default location (bottom-right part stack) of these views. I can understand the desire to get rid of them. If we foresee them perhaps disappearing, it might not be such a good idea to add more functionality to the form headings...

A toolbar button would at least be consistent for all these views.

Patch-set: 1