Fix history view showing only commit IDs sometimes
The history view would sometimes not show the full commit data of a
selected commit but only the first line containing the commit ID.

I finally had a simple reproducible situation, and after spending some
hours debugging our commit parsing and message text formatting without
finding any problem, I finally had the idea to check the viewer itself
and gave is a cyan background. And lo and behold: the viewer was not
resized properly.

So the message text was there all along, but was not shown because the
viewer was too small and showed only the first line!

Fix by forcing a layout of the viewers on text changes.

Also recompute the size when the diff viewer becomes empty.

Bug: 519104
Change-Id: Ibd17fc822e715b0095ed1e87762e92c0d5baaac6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>