Improve asynchronous Eclipse workspace refreshes done by EGit
EGit uses two asynchronous workspace refreshes: one to keep the Eclipse
resource tree in sync with changes made by JGit through the normal
java.io interfaces, which is particularly important when previously
deleted files are restored, and a second one to catch changes made in
a repository outside of Eclipse via git. Both are implemented in the
EGit UI Activator.

The previous implementation refreshed on index changes. This refreshed
far too often -- we actually need to refresh only when something in
the git working tree changes. It also refreshed way too much since it
did full refreshes of all projects from a repository. And it ran all
these refreshes in a single workspace operation holding the workspace
lock.

The net result of all this were very noticeable delays, in particular
on Windows. For instance, staging and unstaging files in the staging
view could get blocked on the refresh for many seconds.

Change [1] makes JGit send events when it modifies anything in the git
working tree. These events contain the precise paths of all modified
or deleted files. With that information, EGit can figure out much more
precisely which parts of the Eclipse workspace to refresh, and since
JGit will never report changes to git-ignored files, EGit will never
even try to refresh those.

This implementation does away with the global workspace lock for
refreshes. IResource.refreshLocal() will acquire an appropriate
scheduling rule on its own.

Tested on Windows with the setup described in bug 500106 (the
unbearable slowness was most pronounced on Windows): with a git-
ignored folder containing 50000 files, operations outside of
that folder (staging/unstaging files, replace with head, ...) are
now fast on Windows, too.

[1] https://git.eclipse.org/r/#/c/100916/

Bug: 500106
Change-Id: Ia97df24e7a3b6fbf89bc55cbcce1a61b8ef8d4fb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>