Update patch set 4
Patch Set 4:

> Why not using the prominent repository label in the upper left of
 > the staging view to do the switching (instead of the new view
 > menu)? That is where I looked first after checking out this change.

That's what patch set 1 did (without the synchronized view switching, though). As mentioned in previous comments, I'm not going to do implement it that way because

1. it's possibly only via hacks.
2. we're thinking about replacing this form heading by something smaller (but still showing which repo is set), so adding more functionality to this heading is probably not a good idea.
3. doing so would leave the history view out in the cold. It doesn't have a form heading and we can't add one since the view is managed by org.eclipse.team, so the only options for that view are a toolbar widget or putting the switcher in the view menu. Either way the switcher would be located elsewhere than in the other views if the other views showed the menu on the form heading.

 > Second comment: For the view menu itself I suggest to use radio
 > button style for the items and to check the current repository.
 > Right now, the menu itself does not help me to identify the current
 > repository.

That's a good idea. I'll do that. Note that if the "current" repo is a bare repo, there will be no checkmark in the menu except in the history view. The history view is the only one that shows bare repos in that menu at all; in the other views, bare repos don't make sense.

Patch-set: 4
Label: Verified=0