Restrict resizing of the "Fetch from Gerrit" Dialog
WizardDialogs in general can be resized to any size; even made
ridiculously small.

Set the shell's minimum size to the size needed to display the
starting page. This allows making the dialog larger, but not smaller
than the starting page.

If this is deemed useful, we could use the new MinimumSizeWizardDialog
as a drop-in replacement for all other WizardDialogs in EGit. It's
also something that platform.ui might want to adopt.

Bug: 519335
Change-Id: I12d967e2e7054304824577cdd0d4a3bd31963d5f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>