Extensible context menu in the rebase interactive view
A double click on a plan item already opened the commit in the commit
viewer. Now add the corresponding action to the view's context menu
so that this functionality is easier to discover.

This also is a nice example of how the new IRepositoryCommit interface
simplifies command management: we just define a new menu contribution
for that context menu, *re-using* the existing handler from the
history view. The new contribution uses its own label since the
accelerator is changed.

Adapt the visibleWhen expressions of the rebase sub-menu items for
the "Team" sub-menu to avoid they appear in the rebase interactive
view.

Moving the existing handler from the "internal.history" package to a
more general place is for now left for a future clean-up commit, as
are more handler unifications and eventually unifying all these
"AbstractXYZHandler" classes that all offer operations to get a
repository and commits from a selection, but that all do it in
slightly different and sometimes even view-specific ways.

Change-Id: I9ae5e709e24bfe3cc938f575df9685d3cfc884bc
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>