Update patch set 3
Patch Set 3:

> I'm not sure. I agree, that it should not matter for a dialog only
 > with a close label. But it is not consistent with e.g. the exit
 > dialog (although this is not the same "kind" of dialog).

I know; I'm not sure myself. But there I have a choice of not exiting, so even if I flipped the toggle, I didn't really make a decision if I then choose not to exit. In Oxygen, the dialog says "Do you want to exit foo?" and the toggle label is "Remember my decision". So if I abandon exiting by closing the dialog in any other way than clicking on the "Exit" button, it feels OK if the toggle is not updated.

Here I only have the choice of closing the dialog. It's a bit strange that if I change the toggle, the state is remembered only if I click on the "Close" button.

Patch-set: 3
Label: Verified=0