Update patch set 3
Patch Set 3:

> Looks like we'll need a way to quickly find files marked assumed
 > unchanged so users can revert the action of marking a file assumed
 > unchanged easily. With git command line this can be done using
 > 
 > $ git ls-files -v | grep '^h'

For the staging view, how about a more ambitious approach: use an undo stack and the standard undo/redo operations? Would that be feasible? Or too expensive because we'd have to remember the list of files operated on? (If so, perhaps limit the undo stack to depth 1?)

Patch-set: 3