Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

When (Compact) Tree View is enabled, the actions are also available on folders and then apply to everything underneath that folder (recursively!). I made a dummy change in EGit UI Activator to have it show up in unstaged, then chose "Untrack" on the folder, and ended up with all files of the EGit UI bundle being removed from the index.

This should behave more like the stage/unstage actions, which apply to all StagingEntry items below a folder. Also note that stage/unstage maintain the selection and expansion state of the tree. If untrack did that, the index deletions would be selected in the staged area, and thus the operation could be quickly undone by unstaging the staged index deletions.

Patch-set: 3
Label: Code-Review=-1