Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

refresh is not triggered for non-workspace resources, e.g. open parent pom.xml of jgit repository in text editor and modify it, save the modification. Then run reset --hard or stash, the pom.xml isn't refreshed, neither in staging or repositories view neither in the text editor. When clicking in the text editor Eclipse recognizes that it's out of sync with filesystem and shows a dialog asking if it should be refreshed

Patch-set: 2
Label: Code-Review=-1