Update patch set 2
Patch Set 2:

(1 comment)

> refresh is not triggered for non-workspace resources, e.g.
 > open parent pom.xml of jgit repository in text editor and
 > modify it, save the modification. Then run reset --hard or
 > stash, the pom.xml isn't refreshed, neither in staging or
 > repositories view neither in the text editor. When clicking
 > in the text editor Eclipse recognizes that it's out of sync
 > with filesystem and shows a dialog asking if it should be
 > refreshed

That's true, but unrelated to this change. Here I'm changing the way projects in the workspace are refreshed, and a file not in the workspace cannot possibly be affected by any change done here. In fact, the behavior you describe also occurs without this change in current EGit 4.8.0.

In staging view we do trigger explicit refreshes for such cases in ReplaceAction. We might consider doing that here, too. I.e., do a workspace refresh but additionally trigger an index diff update for
files outside of the workspace. But I'd prefer to do so in a follow-up commit. In this commit I'd like to concentrate on re-doing the workspace refresh logic only. Adding logic to better deal with non-workspace files is a separate issue.

Patch-set: 2
Label: Verified=0