Update patch set 5
Patch Set 5:

> Thanks a lot, Karsten! Works well now with Untrack.
 > 
 > I must admit that I don't quite understand how Assume Unchanged/No
 > Assume Unchanged are supposed to work; I have never used those. I
 > notice that Assume Unchanged does correctly remove a changed file
 > from the unstaged view, but when I then do a "No Assume Unchanged"
 > on that file in the package explorer, it doesn't re-appear as
 > modified in the staging view. It needs a null edit to re-appear...
 > That, however, is unrelated to this change here.

The assume unchanged flag tells git to ignore the status of the flagged file so (j)git assumes it's unchanged.

I tried this on a file which is in the workspace (in an imported open project) and it works for me, "assume unchanged" removes a modified file from staging view and "no assume unchanged" brings it back.

 > I consider adding an undo mechanism, at least for the last command,
 > as beyond the scope of this change. So from my point of view, this
 > change is good enough.

Patch-set: 5