Add OperationLogger.java and log the operations that change HEAD
You can see logs of operations that change HEAD successfully with
reflog. However, reflog is not suitable to log failed operations
such as "failed to checkout because of non-commited files" or
"merge conflict occured". This commit adds start, end, and error log
to operations that change HEAD.

When an organization is going to change operation of SCM from SVN or
in-house-VCS-based operation to Git-based operation, sometimes they
have to show the effect of change in SCM operations. In this situation,
they have to measure workload of version control with Git and make a
comparison with one in SVN or in-house-VCS-based operation. To do this
measurement, they have to trace operation of each developer on the Git
client and aggregate them.

Additionally, when an organization is going to train the developers who
is new to DVCS, the developers frequentry got stuck with push failures,
marge conflicts, or checkout conflicts. In this situation, they need
the way to analyze cause of the errors later, not immediately after the
errors had occured.

These trace logs are useful for such situations. Based on these use
cases, it is reasonable to limit the target of logging to the command
which communicate with remotes (push, pull, etc.), or change HEAD
(commit, merge, checkout, etc.).

Change-Id: I46dd57fed0ff5f844f0f44eadde2be1b3e6af2dd
Signed-off-by: SATO Yusuke <yusuke.sato.zz@gmail.com>