Add OperationLogger.java and log the operations that change HEAD
You can see logs of operations that change HEAD successfully with
reflog. However, reflog is not suitable to log failed operations
such as "failed to checkout because of non-commited files" or
"merge conflict occured". This commit adds start, end, and error log
to operations that change HEAD. These logs are useful for diagnosis
of problems or training of git operations. 

Change-Id: I46dd57fed0ff5f844f0f44eadde2be1b3e6af2dd
Signed-off-by: SATO Yusuke <yusuke.sato.zz@gmail.com>