Update patch set 1
Patch Set 1:

An additional improvement on this general mechanism in EGit might the following:

As is, EGit will simply log warnings to the Eclipse error log. A user may ignore these warnings. We could add a periodic background job that checks if there have been new such warnings since the last time it ran, and nag the user about it if so. If present, this nagging could perhaps be done via a Mylyn notification (the automatically disappearing info box that appears bottom right on the screen in newer Eclipses), otherwise we might consider that background job return a failure status. Would need to have a preference to set the nag interval (default maybe 5 minutes; user should be able to suppress nagging altogether), and maybe should nag only if the Error log is not set to activate on new entries (because if so, the user will be aware of the warnings already).

What do you think?

Patch-set: 1