Update patch set 2
Patch Set 2: Code-Review-1

The Pattern javadoc explicitly says that \v and \h are not supported character classes. I know that the implementation accepts them neverless.

To make this a bit more fault tolerant I would actually add \s to the pattern, since \v and \h may or may not include \s, given this uncertain documentation. Or is that a bad idea?

Patch-set: 2