Update patch set 2
Patch Set 2: Code-Review+1

When the second line is non-empty we show a warning that the second line should be blank. But with this change we strip the leading blank line effectively moving up the commit message by one line. Though if more leading lines are empty we don't show a warning. If first 2 lines are empty and commit message starts on 3rd line and also has text on the 4th line we don't show this warning but after stripping the leading whitespace we end up with a commit message this warning wants to prevent. 

This looks a bit confusing.

Patch-set: 2