Update patch set 2
Patch Set 2:

> When the second line is non-empty we show a warning that the second
 > line should be blank. But with this change we strip the leading
 > blank line effectively moving up the commit message by one line.
 > Though if more leading lines are empty we don't show a warning. If
 > first 2 lines are empty and commit message starts on 3rd line and
 > also has text on the 4th line we don't show this warning but after
 > stripping the leading whitespace we end up with a commit message
 > this warning wants to prevent.
 > 
 > This looks a bit confusing.

Yes, the warning itself is a bit confusing. This change here was supposed to produce a more meaningful commit message if the user chooses to ignore the warning (leaves one leading empty line) or if there is no warning at all (two or more leading empty lines, or warning switched off). A commit message with an initial leading line just looks bad in Gerrit, in git log when only the short message is shown, and I guess also on Github.

Improving the warning(s) is not the topic of this change. That would be a different task, and yes, we might also warn about leading whitespace including empty lines, and I'd also prefer to see warnings/errors as markers on the fields they relate to over a single dynamic label. But that's quite a bit of work.

Patch-set: 2
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0