Update patch set 1
Patch Set 1:

> Thank you for your advice, I would like to change way to implement.
 > Should I move to the Bug 349551 to discussion forthe specifications
 > of the feature?

No, unless you plan to log into an Eclipse console view. If that is the intent, I could push my prototype for this that I mentioned in that bug report.

But for problem diagnosis and git training, the textual trace file produced by the Eclipse tracing infrastructure may be good enough.

Patch-set: 1