Trim history search input
It is very easy to copy an SHA1 including some whitespace into the
search field, as I accidentally did today. Therefore trimming seems
useful here.

We could restrict the trimming to SHA1 search (instead of all types of
searches) if committers feel that it is an important use case to be able
to search for terms starting/ending with whitespace.

Change-Id: I875530b45c5745a8fe9d6d5972bfc4f357391303
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>