GitHistoryPage: search widget must not react to font changes
The search widget in the page's toolbar must _not_ honor changes
to the JFace dialog font. Otherwise the field becomes unusable
when the user increases the font size.

The problem has several causes, all beyond the control of EGit:

* Bug 521401: at least on Mac, the Text widget itself does not
  adjust its height when the font size changes.
* Bug 183003: even if it did, the Toolbar it is contained in
  does not properly respect the height of ControlContributions
  and would thus not adjust.
* Even if both were true, I'm not sure the CTabFolder containing
  the history view and thus that toolbar would deal properly with
  a toolbar that was higher than the tab height.
  
So it's best to simply use the default font. Font size changes in
the preferences will be ignored, but at least the field will remain
useable.

Change-Id: I96d7b7b5ce714d5bb8573abc58a776b33a37e85f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>