Set window title in 'Init Git Flow' dialog
All other dialogs in the gitflow plugin set the shell title correctly,
as far as I inspected the code.

Additionally I added an ellipsis to the "Init Git Flow" menu, since that
one requires additional input from the user. There seems to be no
corresponding test to be updated. Only the Eclipse wiki could be updated
for the help generation when the change gets merged.

Change-Id: Id95c327ff6e85811ec738ba8a51ce1c64260e440
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>