Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

When running this (on Windows), I see a minor functional difference: When I hover one of the new commands, I get an additional tooltip with the command description after a second. The tooltip is not the identical string as the menu item label, e.g. the menu item "Rebase HEAD on" only shows the tooltip "Rebase".
That tooltip is NOT shown for the "Copy Commit Id", "Add Task Content", "Open Corresponding Task" items at the bottom.

Patch-set: 3
Label: Code-Review=-1