Partially revert "Remove repository from ListRemoteOperation"
This reverts commit f5264cfabfb2f921304536f4471bff469d9ba546. The
repository, if available, provides additional context for running
the ls-remote operation, such as http.* configuration settings in
the repository config. Thus when we do have a repository, we should
use it.

Keep the good bits from commit f5264cf: add a new constructor to
ListRemoteOperation that can work without repository, and use it
in SourceBranchPage instead of creating an empty dummy repository
(which may fail) there. Same in LocalRepositoryTestCase.

Additionally remove the now unused message about not being able to
create a temporary repository.

Change-Id: I7128b3815d8b564a59fbe2a3bb4a21270b221fc4
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>