Update patch set 2
Patch Set 2:

Interesting build failure. It was trying to get org.eclipse.jgit.junit_4.9.0.201709041836.jar from the JGit 4.9.0-SNAPSHOT update site. But that wasn't present there; instead there were only JGit bundles with a timestamp of 201709041847. The two timestamps correspond to JGit builds https://ci.eclipse.org/jgit/job/jgit/1015/ and https://ci.eclipse.org/jgit/job/jgit/1016/ .

So it looks our build happened to read the 1836 artifacts.xml.xz and contents.xml.xz, but then the JGit build overwrote the repo with the 1847 version, and our build then of course didn't find the bundles it wanted.

That's the first time in more than two years that I've seen this happen. Why does this not happen more often? I find it hard to believe that this should have been the first time that an egit-gerrit build ran concurrently with a jgit (nightly) build. Has this something to do with the migration to Jenkins?

Patch-set: 2